requirements.txt
alembic==1.7.5
apispec==5.1.1
asttokens==2.0.5
autopep8==1.6.0
backcall==0.2.0
bcrypt==3.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
decorator==5.1.1
executing==0.8.2
Flask==2.0.2
flask-apispec==0.11.0
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipython==8.0.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
loguru==0.5.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
passlib==1.7.4
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.1
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.29
stack-data==0.1.3
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing-extensions==4.0.1
wcwidth==0.2.5
webargs==8.1.0
Werkzeug==2.0.2
zipp==3.7.0
